\(latex\)-gram frequency tables

Get \(latex\)-gram frequency tables from a training corpus.

get_kgram_freqs(text, dict, N)

Arguments

text
a character vector. The training corpus from which to extract k-gram frequencies.

dict
a character vector. The language model fixed dictionary (see details), sorted by word frequency.

N
a length one integer. The maximum order of \(latex\)-grams for which frequencies are sought.

Value

A kgram_freqs object, containing the \(latex\)-gram frequency tables for \(latex\).

Details

This function extracts all k-gram frequency tables from a text corpus up to a specified \(latex\)-gram order \(latex\). These are the building blocks to train any \(latex\)-gram model.

get_kgram_freqs employs a fixed (user specified) dictionary; any out-of-vocabulary word gets effectively replaced by an "unknown word" token.

The return value is a "kgram_freqs" object, i.e. a list containing:

The highest order of \(latex\)-grams, \(latex\).

The reference dictionary, sorted by word frequency. This can be obtained using get_word_freqs.

A list of n tibbles, storing frequency counts for each \(latex\)-gram observed in the training corpus, for \(latex\). In these tibbles, words are represented by integer numbers corresponding to their position in the reference dictionary. The special codes 0, length(dictionary)+1 and length(dictionary)+2 correspond to the "Begin-Of-Sentence", "End-Of-Sentence" and "Unknown word" tokens, respectively.

See also

Examples

# Obtain k-gram frequency table from corpus
## Get k-gram frequencies, up to k = N = 3.
freqs <- get_kgram_freqs(twitter_train, twitter_dict, N = 3)
## Print result
freqs

#> k-gram frequency table 
#> 
#> Order (N): 3 
#> Dictionary size: 1000  words
#> 
#> # of unique 1-grams: 1002 
#> # of unique 2-grams: 102627 
#> # of unique 3-grams: 410732 
#> 
#> Object size: 13 Mb 
#> 
#> See ?get_kgram_freqs for help.