k-gram frequency tables

Get k-gram frequency tables from a training corpus.

get_kgram_freqs(text, N, dict, .preprocess = preprocess, EOS = ".?!:;")

get_kgram_freqs_fast(
  text,
  N,
  dict,
  erase = "[^.?!:;'\\w\\s]",
  lower_case = TRUE,
  EOS = ".?!:;"
)

Arguments

text
a character vector. The training corpus from which to extract k-gram frequencies.

N
a length one integer. The maximum order of k-grams for which frequencies are sought.

dict
either a character vector, or a length one integer/numeric. The language model fixed dictionary (see details), sorted by word frequency. If numeric, the dictionary is obtained from the training corpus using the dict most frequent words.

.preprocess
a function to apply before k-gram tokenization.

EOS
a length one character vector listing all (single character) end-of-sentence tokens.

erase
a length one character vector. Regular expression matching parts of text to be erased from input. The default removes anything not alphanumeric, white space, apostrophes or punctuation characters (i.e. ".?!:;").

lower_case
a length one logical vector. If TRUE, puts everything to lower case.

Value

A kgram_freqs object, containing the k-gram frequency tables for k = 1, 2, ..., N.

Details

These functions extract all k-gram frequency tables from a text corpus up to a specified k-gram order N. These are the building blocks to train any N-gram model.

The optimized version get_kgram_freqs_fast(erase = x, lower_case = y) is equivalent to get_kgram_freqs(.preprocess = preprocess(erase = x, lower_case = y)), but more efficient (both from the speed and memory point of view).

get_kgram_freqs and get_kgram_freqs_fast employ a fixed (user specified) dictionary; any out-of-vocabulary word gets effectively replaced by an "unknown word" token.

The return value is a "kgram_freqs" object, i.e. a list containing:

The highest order of N-grams, N.

The reference dictionary, sorted by word frequency. This can be obtained using get_word_freqs.

A list of n tibbles, storing frequency counts for each k-gram observed in the training corpus, for k = 1, 2, ..., N. In these tibbles, words are represented by integer numbers corresponding to their position in the reference dictionary. The special codes 0, length(dictionary)+1 and length(dictionary)+2 correspond to the "Begin-Of-Sentence", "End-Of-Sentence" and "Unknown word" tokens, respectively.

The function used for text preprocessing (i.e. the .preprocess argument)

A length one character vector listing all (single character) end-of-sentence tokens. (i.e. EOS argument)

The .preprocess argument of get_kgram_freqs allows the user to employ a custom corpus preprocessing function.

The algorithm for k-gram tokenization considers anything separated by (any number of) white spaces (i.e. " ") as a single word. Sentences are split according to end-of-sentence (single character) tokens, as specified by the EOS argument. Additionally text belonging to different entries of the preprocessed input vector which are understood to belong to different sentences.

See also

Author

Valerio Gherardi

Examples

# \donttest{
# Obtain k-gram frequency table from corpus
## Get k-gram frequencies, up to k = N = 3.
freqs <- get_kgram_freqs(twitter_train, N = 3, dict = twitter_dict)
## Print result
freqs

#> k-gram frequency table 
#> 
#> Order (N): 3 
#> Dictionary size: 1000  words
#> 
#> # of unique 1-grams: 1002 
#> # of unique 2-grams: 85677 
#> # of unique 3-grams: 318410 
#> 
#> Object size: 5.9 Mb 
#> 
#> See ?get_kgram_freqs for help. 

# }
# \donttest{
# Obtain k-gram frequency table from corpus
## Get k-gram frequencies, up to k = N = 3.
freqs <- get_kgram_freqs_fast(twitter_train, N = 3, dict = twitter_dict)
## Print result
freqs

#> k-gram frequency table 
#> 
#> Order (N): 3 
#> Dictionary size: 1000  words
#> 
#> # of unique 1-grams: 1002 
#> # of unique 2-grams: 85677 
#> # of unique 3-grams: 318410 
#> 
#> Object size: 5.9 Mb 
#> 
#> See ?get_kgram_freqs for help. 

# }